Made form useable down to 400px.
Edited css so that the content would be usable all of the way down to 400px.
To Do: Fix the way the dynamic rows work so that they are positioned in a manner that makes sense and so that the buttons are responsive under 400px.
Also need to add animations and fix the main table on the front page and add nav menu to smaller sizes.